Isolate EXTERNALLY-MANAGED logic better
